pyproject.toml
[project]
name = "mnist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.2.4",
    "omegaconf>=2.3.0",
    "pytorch-lightning>=2.5.1",
    "tensorboard>=2.19.0",
    "torchvision>=0.21.0",
]
